+++ routes/hotel/login.js
@@ -242,7 +242,7 @@ route.post('/verify/:id', async (req, res) => {
         if (results[0].pin === req.body.pin) {
 
           if (results) {
-            Verification.deleteOne({ id })
+            Verification.deleteOne({ userId:id })
               .then(() => {
                 return res.send({
                   status: 'Success',
